Use useInterval for fetching workspace url
